requirements.txt
python-telegram-bot == 13.7
pyowm == 3.2.0